+++ CHANGELOG.md
@@ -4,6 +4,9 @@ This project adheres to [Semantic Versioning](http://semver.org/).
 This CHANGELOG follows the format listed at [Keep A Changelog](http://keepachangelog.com/)
 
 ## [Unreleased]
+### Added
+- Add support for making multiple requests, with a threshold for success
+
 ## [1.1.0]
 ### Added
 - Added option for DNS server port (@apriljo)